requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.15
django-formtools==2.5.1
django-otp==1.5.2
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==28.4.1
frozenlist==1.4.1
idna==3.8
inflection==0.5.1
kombu==5.4.0
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.1
phonenumbers==8.13.44
pillow==10.4.0
prompt_toolkit==3.0.47
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
requests==2.32.3
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.1
twilio==9.2.4
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
